split models into separate files
