+++ search-api/search_api/models/base.py
@@ -1,0 +1,44 @@
+# Copyright © 2020 Province of British Columbia
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from decimal import Decimal
+
+import flask.json
+from flask_sqlalchemy import SQLAlchemy
+
+
+class MyJSONEncoder(flask.json.JSONEncoder):
+
+    def default(self, obj):
+        if isinstance(obj, Decimal):
+            # Convert decimal instances to strings.
+            return str(obj)
+        return super(MyJSONEncoder, self).default(obj)
+
+
+flask.json_encoder = MyJSONEncoder
+db = SQLAlchemy()
+
+
+class BaseModel(db.Model):
+    __abstract__ = True
+
+    def as_dict(self):
+        d = {}
+        for c in self.__table__.columns:
+            d[c.name] = getattr(self, c.name)
+            if type(d[c.name]) is Decimal:
+                d[c.name] = int(d[c.name])
+        return d
