+++ search-api/search_api/models/corp_op_state.py
@@ -11,3 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from search_api.models.base import BaseModel, db
+
+
+class CorpOpState(BaseModel):
+    # A lookup table of states a corporation can be in.
+    __tablename__ = 'corp_op_state'
+    """
+    state_typ_cd       CHAR        3     31
+    op_state_typ_cd    CHAR        3     31
+    short_desc         VARCHAR2    15    31
+    full_desc          VARCHAR2    40    31
+    """
+
+    state_typ_cd = db.Column(db.String(3), primary_key=True)
+    op_state_typ_cd = db.Column(db.String(3))
+    short_desc = db.Column(db.String(15))
+    full_desc = db.Column(db.String(40))
