+++ search-api/search_api/models/office_type.py
@@ -11,3 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from search_api.models.base import BaseModel, db
+
+
+class OfficeType(BaseModel):
+    __tablename__ = "office_type"
+    """
+    office_typ_cd    CHAR        2     9
+    short_desc       VARCHAR2    15    9
+    full_desc        VARCHAR2    40    9
+    """
+
+    office_typ_cd = db.Column(db.String(2), primary_key=True)
+    short_desc = db.Column(db.String(15))
+    full_desc = db.Column(db.String(40))
